REFACTOR: 클래스 통합, 메서드 이동 및 분리
View 메서드를 하나의 클래스로 통합했습니다.
애플리케이션의 게임실행 메서드를 Game 클래스로 이동했습니다.
게임을 계속할지 확인하는 과정을 별도의 메서드로 분리했습니다.
