--- src/main/java/baseball/view/InputView.java
@@ -1,12 +1,0 @@
-package baseball.view;
-
-public class InputView {
-
-    public static void askNumber() {
-        System.out.print("숫자를 입력해주세요 : ");
-    }
-
-    public static void askOneOrTwo() {
-        System.out.println("게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.");
-    }
-}
