+++ src/main/java/baseball/view/View.java
@@ -1,6 +1,14 @@
 package baseball.view;
 
-public class OutputView {
+public class View {
+
+    public static void askNumber() {
+        System.out.print("숫자를 입력해주세요 : ");
+    }
+
+    public static void askOneOrTwo() {
+        System.out.println("게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.");
+    }
 
     public static void showOpening() {
         System.out.println("숫자 야구 게임을 시작합니다.");
